requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3  # For faster XML parsing (Yandex API responses)
dnspython==2.4.2  # For email DNS MX validation
pandas==2.1.3
selenium==4.15.2
undetected-chromedriver>=3.5.4  # For TruePeopleSearch CAPTCHA bypass
phonenumbers==8.13.24
google==3.0.0
googlesearch-python==1.2.3
python-dotenv==1.0.0
redis==5.0.1
ratelimit==2.2.1
jinja2==3.1.2
folium==0.15.0
matplotlib==3.8.2
seaborn==0.13.0
colorama==0.4.6
twilio==8.10.1
flask==3.0.0
PySocks==1.7.1  # For SOCKS5 proxy support